发送 Thread 消息时只 loading Thread 的 send button
