--- src/features/Portal/Thread/Body/index.tsx
@@ -1,13 +1,0 @@
-import { Flexbox } from 'react-layout-kit';
-
-import Chat from '../Chat';
-
-const ThreadBody = () => {
-  return (
-    <Flexbox flex={1}>
-      <Chat />
-    </Flexbox>
-  );
-};
-
-export default ThreadBody;
