+++ src/features/Portal/Thread/index.ts
@@ -1,10 +1,10 @@
 import { PortalImpl } from '../type';
-import Body from './Body';
+import Chat from './Chat';
 import Header from './Header';
 import { onClose, useEnable } from './hook';
 
 export const Thread: PortalImpl = {
-  Body,
+  Body: Chat,
   Header,
   Title: () => null,
   onClose,
